Fix formatting for the non-js version
We need a scrollable iframe for the non-js fallback
because it is very long, and we can't use js to fix
the height of the iframe once we have loaded and
found out how tall it needs to be.
However, a scrollable iframe with content which fits
it perfectly will result in scrollbars on some
browsers, which reduces the available space and means
we'd need to scroll on the js version.
Also, changing the scrollable attributes on an iframe
does nothing once the page has loaded, so we can't
just declare the iframe scrollable and turn off
scrolling using js if it's available.

We have to resort to a div containing the iframe,
so that we can replace the innerhtml completely with
a differently declared iframe if js is available.
Seems to work ok though.